First pass of adding OMNeTpp 6 support
Mostly changed use of longValue to intValue in cPars, as the interface
for cPars has changed. Also rectified an issue where an iterator was not
correctly passed to another owner.

+++ policies/DIF/RA/PDUFG/GRE/ClosStatic/GRE_Clos1S.cc
@@ -75,9 +75,9 @@ void GRE_ClosS1::routingUpdated(){}
 
 // Called after initialize
 void GRE_ClosS1::onPolicyInit(){
-    f = par("spineS").longValue();
-    t = par("TEs").longValue();
-    s = par("spines").longValue();
+    f = par("spineS").intValue();
+    t = par("TEs").intValue();
+    s = par("spines").intValue();
 
     //Set Forwarding policy
     fwd = getRINAModule<Clos1 *>(this, 2, {MOD_RELAYANDMUX, MOD_POL_RMT_PDUFWD});
