+++ policies/DIF/RA/PDUFG/LatGenerator/LatGenerator.cc
@@ -44,16 +44,16 @@ bool portMetric::operator < (const portMetric &other) const {
 void LatGenerator::insertedFlow(const Address &addr, const QoSCube &qos, RMTPort * port){
     std::string dst = addr.getIpcAddress().getName();
 
-    unsigned short metric = qos.getDelay()/par("redLinkCost").longValue();
+    unsigned short metric = qos.getDelay()/par("redLinkCost").intValue();
 
-    if(metric < par("minLinkCost").longValue()) {
-        metric = par("minLinkCost").longValue();
-    } else if(metric > par("maxLinkCost").longValue()) {
-        metric = par("maxLinkCost").longValue();
+    if(metric < par("minLinkCost").intValue()) {
+        metric = par("minLinkCost").intValue();
+    } else if(metric > par("maxLinkCost").intValue()) {
+        metric = par("maxLinkCost").intValue();
     }
 
     if(qos.getQosId() == qos.MANAGEMENT.getQosId() || qos.getQosId() == VAL_UNDEF_QOSID) {
-        metric = par("maxLinkCost").longValue();
+        metric = par("maxLinkCost").intValue();
     }
 
     neighbours[dst].insert(portMetric(port, metric));
@@ -79,7 +79,7 @@ void LatGenerator::removedFlow(const Address &addr, const QoSCube& qos, RMTPort 
         neighbours.erase(dst);
         routingUpdated();
     } else {
-        unsigned short min = par("maxLinkCost").longValue();
+        unsigned short min = par("maxLinkCost").intValue();
         for(portMetric mt : neighbours[dst]) {
             if(min >= mt.metric) { min = mt.metric; }
         }
