+++ policies/DIF/RA/PDUFG/LatencySingle1Entry/LatencySingle1Entry.cc
@@ -37,16 +37,16 @@ bool portMetric::operator < (const portMetric &other) const {
 void LatencySingle1Entry::insertedFlow(const Address &addr, const QoSCube &qos, RMTPort * port){
     string dst = addr.getIpcAddress().getName();
 
-    mType metric = qos.getDelay()/par("redLinkCost").longValue();
+    mType metric = qos.getDelay()/par("redLinkCost").intValue();
 
-    if(metric < par("minLinkCost").longValue()) {
-        metric = par("minLinkCost").longValue();
-    } else if(metric > par("maxLinkCost").longValue()) {
-        metric = par("maxLinkCost").longValue();
+    if(metric < par("minLinkCost").intValue()) {
+        metric = par("minLinkCost").intValue();
+    } else if(metric > par("maxLinkCost").intValue()) {
+        metric = par("maxLinkCost").intValue();
     }
 
     if(qos.getQosId() == qos.MANAGEMENT.getQosId() || qos.getQosId() == VAL_UNDEF_QOSID) {
-        metric = par("maxLinkCost").longValue();
+        metric = par("maxLinkCost").intValue();
     }
 
     neighbours[dst].insert(portMetric(port, metric));
@@ -72,7 +72,7 @@ void LatencySingle1Entry::removedFlow(const Address &addr, const QoSCube& qos, R
         neighbours.erase(dst);
         routingUpdated();
     } else {
-        unsigned short min = par("maxLinkCost").longValue();
+        unsigned short min = par("maxLinkCost").intValue();
         for(portMetric mt : neighbours[dst]) {
             if(min >= mt.metric) { min = mt.metric; }
         }
