+++ src/DAF/RIB/DAFRIBdBase.cc
@@ -29,7 +29,7 @@ DAFRIBdBase::~DAFRIBdBase() {
 }
 
 long DAFRIBdBase::getNewInvokeId() {
-    long newinvoke = getParentModule()->getSubmodule(MOD_CDAP)->par(PAR_CURINVOKEID).longValue() + 1;
+    long newinvoke = getParentModule()->getSubmodule(MOD_CDAP)->par(PAR_CURINVOKEID).intValue() + 1;
     getParentModule()->getSubmodule(MOD_CDAP)->par(PAR_CURINVOKEID) = newinvoke;
     return newinvoke;
 }
